WIP GLTFLoader to TS
