+++ src/loaders/tmp/getImageURIMimeType.ts
@@ -1,0 +1,6 @@
+export function getImageURIMimeType(uri) {
+  if (uri.search(/\.jpe?g($|\?)/i) > 0 || uri.search(/^data\:image\/jpeg/) === 0) return 'image/jpeg'
+  if (uri.search(/\.webp($|\?)/i) > 0 || uri.search(/^data\:image\/webp/) === 0) return 'image/webp'
+
+  return 'image/png'
+}
